requirements.txt
pandas>=1.5.0
numpy>=1.20.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
torch>=1.10.0
xanfis>=0.1.0
mealpy>=2.5.0
permetrics>=1.3.0
shap>=0.41.0
PyYAML>=6.0
